commennts

+++ src/com/sophia/PreparationKit/MiniMaxSum.java
@@ -16,9 +16,9 @@ public class MiniMaxSum {
 
         public static void miniMaxSum(List<Integer> arr) {
                 Collections.sort(arr);
-                int minSum = 0;
-                int maxSum = 0;
-                int sum = 0;
+                long minSum = 0;
+                long maxSum = 0;
+                long sum = 0;
                 for(int i = 0;i<arr.size();i++){
                     sum += arr.get(i);
                     minSum = sum - arr.get(arr.size()-1);
